+++ core-ng/src/main/java/core/framework/http/HTTPClientBuilder.java
@@ -96,7 +96,9 @@ public final class HTTPClientBuilder {
                 .callTimeout(callTimeout()) // call timeout is only used as last defense, timeout for complete call includes connect/retry/etc
                 .connectionPool(new ConnectionPool(100, keepAlive.toSeconds(), TimeUnit.SECONDS))
                 .eventListenerFactory(new HTTPEventListenerFactory())
-                .retryOnConnectionFailure(false);   // use RetryInterceptor to record all failures, RetryAndFollowUpInterceptor will always be added to chain by okHTTP, so it still supports followup
+                .retryOnConnectionFailure(false)    // disable all okHTTP builtin retry and followups, those should be handled on application level for traces
+                .followRedirects(false)
+                .followSslRedirects(false);
 
             configureHTTPS(builder);